pyproject.toml
[project]
name = "texor"
version = "0.1.0"
description = ""
authors = [
    {name = "coderTtxi12",email = "oliver.coding8@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "langgraph (>=0.2.74,<0.3.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "tavily-python (>=0.5.1,<0.6.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pypdf (>=5.3.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^25.1.0"
isort = "^6.0.0"

